+++ tests/testthat/test-cv_nested_surv_glmnet_cox.R
@@ -13,7 +13,7 @@ param_list_glmnet <- expand.grid(
 )
 glmnet_bounds <- list(alpha = c(0., 1.))
 optim_args <- list(
-  iters.n = 4L,
+  iters.n = 2L,
   kappa = 3.5,
   acq = "ucb",
   otherHalting = list(timeLimit = 30)
@@ -40,22 +40,30 @@ train_y <- survival::Surv(
 
 fold_list <- splitTools::create_folds(
   y = split_vector,
-  k = 10,
+  k = 5,
   type = "stratified",
   seed = seed
 )
 
+
 test_that(
-  desc = "test cv - surv_glmnet_cox",
+  desc = "test nested cv, bayesian - surv_glmnet_cox",
   code = {
 
-    surv_glmnet_cox_cv <- MLCrossValidation$new(
+    surv_glmnet_cox_cv <- MLNestedCV$new(
       learner = learner,
+      strategy = "bayesian",
       fold_list = fold_list,
+      k_tuning = 3L,
       ncores = 2L,
       seed = seed
     )
-    surv_glmnet_cox_cv$learner_args = list(alpha = 0.8, lambda = 0.002)
+
+    surv_glmnet_cox_cv$parameter_bounds <- glmnet_bounds
+    surv_glmnet_cox_cv$parameter_grid <- param_list_glmnet
+    surv_glmnet_cox_cv$split_type <- "stratified"
+    surv_glmnet_cox_cv$split_vector <- split_vector
+    surv_glmnet_cox_cv$optim_args <- optim_args
 
     # set data
     surv_glmnet_cox_cv$set_data(
@@ -65,7 +73,40 @@ test_that(
 
     cv_results <- surv_glmnet_cox_cv$execute()
     expect_type(cv_results, "list")
-    expect_equal(dim(cv_results), c(10, 3))
+    expect_equal(dim(cv_results), c(5, 3))
     expect_true(inherits(x = surv_glmnet_cox_cv$results, what = "mlexCV"))
   }
 )
+
+
+test_that(
+  desc = "test nested cv, grid - surv_glmnet_cox",
+  code = {
+
+    surv_glmnet_cox_cv <- MLNestedCV$new(
+      learner = learner,
+      strategy = "grid",
+      fold_list = fold_list,
+      k_tuning = 3L,
+      ncores = 2L,
+      seed = seed
+    )
+
+    surv_glmnet_cox_cv$parameter_grid <- param_list_glmnet
+    surv_glmnet_cox_cv$split_type <- "stratified"
+    surv_glmnet_cox_cv$split_vector <- split_vector
+    surv_glmnet_cox_cv$optim_args <- optim_args
+
+    # set data
+    surv_glmnet_cox_cv$set_data(
+      x = train_x,
+      y = train_y
+    )
+
+    cv_results <- surv_glmnet_cox_cv$execute()
+    expect_type(cv_results, "list")
+    expect_equal(dim(cv_results), c(5, 3))
+    expect_true(inherits(x = surv_glmnet_cox_cv$results, what = "mlexCV"))
+  }
+)
+