+++ NEWS.md
@@ -1,3 +1,25 @@
 # mlexperiments NEWS
 
-## Unreleased (2022-09-27)
+## Unreleased (2022-10-01)
+
+#### New features
+
+-   initial commits
+    ([07f39da](https://github.com/kapsner/mlexperiments/tree/07f39da254f6ae51248361e29b3c42888c1f2325))
+
+#### Bug fixes
+
+-   fixed last issue with seed in grid-search
+    ([cc821d6](https://github.com/kapsner/mlexperiments/tree/cc821d6f1c2b4364e19a92d7277fcda5fbfc7c95))
+-   fixed issue related to parallel backend in bayesian tuning
+    ([2afba3f](https://github.com/kapsner/mlexperiments/tree/2afba3ff6c1407d8b87942662bb14e48a4497366))
+-   removed duplicate class names and fixed related issues
+    ([9ca9385](https://github.com/kapsner/mlexperiments/tree/9ca9385167841ccfea490e05f42e94f07f757cac))
+
+#### Refactorings
+
+-   reorgainzed code to make use of more inheritance
+    ([772bba0](https://github.com/kapsner/mlexperiments/tree/772bba0cadd439f65fbf04ede5899948acc5b53c))
+
+Full set of changes:
+[`edb76c3...cc821d6`](https://github.com/kapsner/mlexperiments/compare/edb76c3...cc821d6)